:construction: GI-4 Change int to float casting

+++ src/models/api_responses/dodo/delivery_productivity.py
@@ -7,6 +7,6 @@ __all__ = ('UnitDeliveryProductivityStatisticsReport',)
 
 class UnitDeliveryProductivityStatisticsReport(BaseModel):
     unit_uuid: UUID
-    orders_per_courier_labour_hour_today: int
-    orders_per_courier_labour_hour_week_before: int
+    orders_per_courier_labour_hour_today: float
+    orders_per_courier_labour_hour_week_before: float
     from_week_before_in_percents: int
